Move unsubscribe link generation into own module
It's not just used by the cronjob, but also by the admin tool on
/admin/emails/. Since that is a Next.js page, if the function is
imported from a module that also imports the cronjob l10n module,
it breaks. Thus, I've moved it into its own module.
