+++ src/scripts/cronjobs/monthlyActivityFree.tsx
@@ -3,8 +3,6 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 import { SubscriberRow } from "knex/types/tables";
-import { captureException } from "@sentry/node";
-import { randomBytes } from "crypto";
 import { getFreeSubscribersWaitingForMonthlyEmail } from "../../db/tables/subscribers";
 import { initEmail, sendEmail } from "../../utils/email";
 import { renderEmail } from "../../emails/renderEmail";
@@ -16,12 +14,9 @@ import { getLatestOnerepScanResults } from "../../db/tables/onerep_scans";
 import { getSubscriberBreaches } from "../../app/functions/server/getSubscriberBreaches";
 import { refreshStoredScanResults } from "../../app/functions/server/refreshStoredScanResults";
 import { getSignupLocaleCountry } from "../../emails/functions/getSignupLocaleCountry";
-import {
-  addUnsubscribeTokenForSubscriber,
-  getEmailPreferenceForSubscriber,
-  updateEmailPreferenceForSubscriber,
-} from "../../db/tables/subscriber_email_preferences";
+import { updateEmailPreferenceForSubscriber } from "../../db/tables/subscriber_email_preferences";
 import { logger } from "../../app/functions/server/logging";
+import { getMonthlyActivityFreeUnsubscribeLink } from "../../app/functions/cronjobs/unsubscribeLinks";
 
 void run();
 
@@ -112,41 +107,5 @@ async function sendMonthlyActivityEmail(subscriber: SubscriberRow) {
     logger.error("send_monthly_activity_email_free", {
       exception: e,
     });
-  }
-}
-
-export async function getMonthlyActivityFreeUnsubscribeLink(
-  subscriber: SubscriberRow,
-) {
-  try {
-    const newUnsubToken = randomBytes(64).toString("hex");
-    let sub;
-    const getRes = await getEmailPreferenceForSubscriber(subscriber.id);
-    if (getRes.unsubscribe_token) {
-      // if record has been created and the token exists, return the token
-      return `${process.env.SERVER_URL}/unsubscribe-email/monthly-report-free?token=${getRes.unsubscribe_token}`;
-    } else if (
-      !getRes.monthly_monitor_report_free_at &&
-      !getRes.unsubscribe_token
-    ) {
-      // if record in the new table has not been created
-      sub = await addUnsubscribeTokenForSubscriber(
-        subscriber.id,
-        newUnsubToken,
-      );
-    } else {
-      // if record already exists, but token doesn't exist, add the token
-      sub = await updateEmailPreferenceForSubscriber(subscriber.id, true, {
-        unsubscribe_token: newUnsubToken,
-      });
-    }
-
-    return `${process.env.SERVER_URL}/unsubscribe-email/monthly-report-free?token=${sub.unsubscribe_token}`;
-  } catch (e) {
-    logger.error("generate_unsubscribe_link", {
-      exception: e,
-    });
-    captureException(e);
-    return null;
   }
 }
